Split out book tests into test module with multiple files
