+++ derrida/books/tests/test_models.py
@@ -1,49 +1,25 @@
 # -*- coding: utf-8 -*-
-import logging
-from collections import defaultdict
-import csv
+import os
+
 from django.contrib.auth import get_user_model
-from django.core.management import call_command
-from django.http import JsonResponse
 from django.utils.safestring import mark_safe
 from django.test import TestCase
+from django.urls import reverse
 import pytest
-
-try:
-    # django 1.10
-    from django.urls import reverse
-except ImportError:
-    from django.core.urlresolvers import reverse
-import json
-from unittest.mock import patch
-import os
-import re
-from io import StringIO
-
-# Import commmand
-from .management.commands import import_zotero
 
 # All the models
 from derrida.places.models import Place
 from derrida.people.models import Person
 # Common models between projects and associated new types
-from .models import AssociatedBook, Book, Catalogue, Creator, CreatorType, \
+from derrida.books.models import AssociatedBook, Book, Catalogue, Creator, CreatorType, \
     ItemType, Publisher, OwningInstitution
 # Citationality extensions
-from .models import DerridaWork, DerridaWorkBook, Reference, ReferenceType, \
+from derrida.books.models import DerridaWork, DerridaWorkBook, Reference, ReferenceType, \
     Work
-# data migration from book to work
-from derrida.books import migration_utils
-
-# pandas
-from pandas import DataFrame
 
 # User model
 User = get_user_model()
 
-
-FIXTURE_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)),
-    'fixtures')
 
 class TestOwningInstitution(TestCase):
     fixtures = ['sample_book_data.json']
@@ -505,283 +481,4 @@ class TestReference(TestCase):
         references[0].delete()
         references = Reference.objects.filter(book=bk1)
         assert references.count() == 1
-
-
-class TestImportZotero(TestCase):
-
-    test_csv = os.path.join(FIXTURE_DIR, 'test_zotero_data.csv')
-
-    def setUp(self):
-
-        self.cmd = import_zotero.Command()
-        self.cmd.stdout = StringIO()
-        self.cmd.stats = defaultdict(int)
-        self.cmd.dud_code_list = []
-
-        # Kludge to get dummy values in the command we're using to test for
-        # API lookups, since we're already hacking it a bit.
-        # TODO: Figure out how to mock this instead
-        def dummy_viaf(*args, **kwargs):
-            return 'http://totallyviaf.org/viaf/00001/'
-        def dummy_geonames(*args, **kwargs):
-            return {
-                'latitude': 0,
-                'longitude': 0,
-                'geonames_id': 'http://notgeonames/0001/'
-            }
-
-        self.cmd.viaf_lookup = dummy_viaf
-        self.cmd.geonames_lookup = dummy_geonames
-
-        princeton, created = Place.objects.get_or_create(
-            name='Princeton, NJ',
-            **(self.cmd.geonames_lookup('Princeton, NJ'))
-        )
-        OwningInstitution.objects.get_or_create(
-            short_name='PUL',
-            contact_info='http://library.princeton.edu/help/contact-us',
-            place=princeton
-        )
-
-    def test_run(self):
-            out = StringIO()
-            # pass the modified self.cmd object
-            call_command(self.cmd, self.test_csv, stdout=out)
-            output = out.getvalue()
-            assert '5 books' in output
-            assert '2 places' in output
-            assert '7 people' in output
-            assert '2 publishers' in output
-            assert '23 references from tags' in output
-
-            assert '0 errors' in output
-            assert 'following tags looked suspicious' not in output
-            assert '0 references with unset tags'
-
-    def test_clean_data(self):
-        data = self.cmd.clean_data(self.test_csv)
-        assert isinstance(data, DataFrame)
-        wanted_columns = [
-            'Key',
-            'Item Type',
-            'Date',
-            'Author',
-            'Title',
-            'Translator',
-            'Editor',
-            'Series Editor',
-            'Url',
-            'Publisher',
-            'Place',
-            'Language',
-            'Extra',
-            'Abstract Note',
-            'Notes',
-            'Manual Tags',
-            'Pages',
-            'Publication Title',
-            ]
-
-        for column in wanted_columns:
-            assert column in data
-
-    def test_parse_ref_tag(self):
-        pub, created = Publisher.objects.get_or_create(name='Pub Lee')
-        pub_place, created = Place.objects.get_or_create(
-            name='Printington',
-            geonames_id=4567,
-            latitude=0,
-            longitude=0
-        )
-        book = ItemType.objects.get(name='Book')
-        bk, created = Book.objects.get_or_create(
-            primary_title='Some rambling long old title',
-            item_type=book,
-            short_title='Some rambling',
-            original_pub_info='foo',
-            publisher=pub,
-            pub_place=pub_place,
-            work_year=1823
-        )
-        dg = DerridaWork.objects.get(short_title='De la grammatologie')
-        # Get a tag that's a full format
-        full_tag = 'dg109cF10sU'
-
-        self.cmd.parse_ref_tag(full_tag, bk)
-        reference = Reference.objects.get(book=bk)
-        assert reference.derridawork == dg
-        assert reference.derridawork_page == '109'
-        assert reference.derridawork_pageloc == 'c'
-        assert reference.reference_type.name == 'Footnote'
-        # page + sequelae
-        assert reference.book_page == '10s'
-        # Unknown if it's annotated or not, s
-        self.assertFalse(bk.is_extant)
-        self.assertFalse(bk.is_annotated)
-        # Make sure two can be added by the short_description
-        full_tag = 'dg111cF10sU'
-        self.cmd.parse_ref_tag(full_tag, bk)
-        references = Reference.objects.filter(book=bk)
-        assert len(references) == 2
-        for reference in references:
-            reference.delete()
-
-        # Now play with tag types
-        full_tag = 'dg112eF(IV)sU'
-        self.cmd.parse_ref_tag(full_tag, bk)
-        reference = Reference.objects.get(book=bk)
-        assert reference.book_page == '(IV)s'
-        reference.delete()
-
-        full_tag = 'dg112eQ(IV)sU'
-        self.cmd.parse_ref_tag(full_tag, bk)
-        reference = Reference.objects.get(book=bk)
-        assert reference.reference_type.name == 'Quotation'
-        reference.delete()
-
-        # provisional tag has no book page
-        full_tag = 'dg112eQ___'
-        self.cmd.parse_ref_tag(full_tag, bk)
-        reference = Reference.objects.get(book=bk)
-        assert reference.reference_type.name == 'Quotation'
-        assert reference.book_page == ''
-        reference.delete()
-
-    def test_create_book(self):
-        data = self.cmd.clean_data(self.test_csv)
-        rows = []
-        for index, row in data.iterrows():
-            rows.append(row)
-        de_gram = rows[0]
-
-        # Make the first work - journal article
-        article = ItemType.objects.get(name='Journal Article')
-        self.cmd.create_book(de_gram)
-        book = Book.objects.get(short_title__contains='De la')
-        assert book.primary_title == de_gram['Title']
-        assert book.short_title == ' '.join(de_gram['Title'].split()[0:4])
-        # No publisher so should be None
-        self.assertFalse(book.original_pub_info)
-        assert book.page_range ==  de_gram['Pages']
-        # Reflecting no finding aid and annotations need to be set
-        # manually, defaulting to false
-        self.assertFalse(book.is_annotated)
-        self.assertFalse(book.is_extant)
-        assert book.item_type == article
-        assert book.copyright_year == 1965
-        # Make sure the author was associated
-        assert book.authors().count() == 1
-        derrida = 'Derrida, Jacques'
-        assert book.authors().first().person.authorized_name == derrida
-        # Make sure a catalogue exists for Princeton
-        catalogue = Catalogue.objects.get(book=book)
-        assert catalogue.institution.short_name == 'PUL'
-        # There should be one tag
-        references = Reference.objects.filter(book=book)
-        assert references.count() == 1
-        # Check a book with notes fields of some sorts
-        idees = rows[4]
-        self.cmd.create_book(idees)
-        book = Book.objects.get(short_title__contains='Idées')
-        assert book.notes
-        # Only one of three note fields set
-        assert book.notes == idees['Abstract Note']
-        # Let's fix that
-        book.delete()
-        idees['Notes'] = 'Notes field'
-        idees['Extra'] = 'Extra field'
-        self.cmd.create_book(idees)
-        book = Book.objects.get(short_title__contains='Idées')
-        # Now all three should exist on their own separate line
-        assert book.notes == '\n'.join([idees['Notes'],
-                                      idees['Extra'],
-                                      idees['Abstract Note']])
-
-        # Faked New York to check place setting functioning correctly
-        print(book.pub_place)
-        assert book.pub_place == Place.objects.get(name='New York')
-
-class TestPubAutocomplete(TestCase):
-
-    def setUp(self):
-        user = User.objects.create_superuser(
-            username='test',
-            password='secret',
-            email='foo@bar.com'
-        )
-        pub = Publisher.objects.create(name='Printing',)
-
-    def test_view_behavior(self):
-        # No login for anonymous user
-        response = self.client.get(reverse('books:publisher-autocomplete'))
-        assert response.status_code == 302
-
-        # Get a response as a staff user
-        self.client.login(username='test', password='secret')
-        response = self.client.get(
-            reverse('books:publisher-autocomplete'),
-            params = {'q': 'Print'}
-        )
-        assert response.status_code == 200
-        data = json.loads(response.content.decode('utf-8'))
-        assert 'results' in data
-        assert data['results'][0]['text'] == 'Printing'
-
-
-# migration utilities for splitting books out into works/instances
-
-def test_cleaned_title():
-    assert migration_utils.cleaned_title('foo') == 'foo'
-    assert migration_utils.cleaned_title('essais (manuscrit)') == 'essais'
-    assert migration_utils.cleaned_title('Emile [some edition]') == 'Emile'
-    assert migration_utils.cleaned_title("Émile ou de l'éducation") == \
-        "Emile ou de l'education"
-
-
-def test_parse_page_range():
-    assert migration_utils.parse_page_range('3-4') == ['3', '4']
-    assert migration_utils.parse_page_range('123-456') == ['123', '456']
-    assert migration_utils.parse_page_range('I-LII') == ['I', 'LII']
-    assert migration_utils.parse_page_range('123') == ['123', '123']
-
-    with pytest.raises(Exception):
-        migration_utils.parse_page_range('Jan-32')
-
-
-@pytest.mark.django_db
-def test_belongs_to_work():
-    # create book and work to test comparison
-    author = Person.objects.create(authorized_name='An Author')
-    author2 = Person.objects.create(authorized_name='Another Author')
-    item_type = ItemType.objects.all().first()
-    CreatorType.objects.create(name='Author')
-    book = Book.objects.create(primary_title='Essais',
-        item_type=item_type)
-
-    work = Work.objects.create(primary_title='Essais')
-
-    # matching titles, no authors
-    assert migration_utils.belongs_to_work(book, work)
-    # case difference in titles
-    book.primary_title = 'essais'
-    assert migration_utils.belongs_to_work(book, work)
-    # extra content in title
-    book.primary_title = 'Essais (manuscrit)'
-    assert migration_utils.belongs_to_work(book, work)
-    book.primary_title = 'Essais [later edition]'
-    assert migration_utils.belongs_to_work(book, work)
-
-    # author mismatch
-    work.authors.add(author)
-    assert not migration_utils.belongs_to_work(book, work)
-    # one author matches
-    book.add_author(author)
-    assert migration_utils.belongs_to_work(book, work)
-    # author work has more authors
-    work.authors.add(author2)
-    assert not migration_utils.belongs_to_work(book, work)
-    book.add_author(author2)
-    # author match with two authors
-    assert migration_utils.belongs_to_work(book, work)
-
