+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE=derrida.settings
 # look for tests in standard django test location
-python_files = "derrida/**/tests.py"
+python_files = "derrida/**/tests.py" "derrida/**/tests/*.py"
